chore: use the same opts for bundler
